requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cleo==2.1.0
crashtest==0.4.1
cryptography==45.0.6
distlib==0.4.0
dulwich==0.22.8
et_xmlfile==2.0.0
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.6.3
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
keyring==25.6.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.4
openpyxl==3.1.5
packaging==25.0
pbs-installer==2025.8.28
pdf2image==1.17.0
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry==2.1.4
poetry-core==2.1.3
propcache==0.3.2
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dotenv==1.1.1
RapidFuzz==3.14.0
requests==2.32.4
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
tomlkit==0.13.3
trove-classifiers==2025.8.26.11
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.32.0
xlrd==2.0.2
yarl==1.20.1
zstandard==0.24.0